requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
mypy==1.17.0
mypy_extensions==1.1.0
oauthlib==3.3.1
pathspec==0.12.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.1
python-json-logger==3.3.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
